+++ mitra_activitypub/src/builders/note.rs
@@ -554,7 +554,9 @@ mod tests {
                 "https://w3id.org/security/data-integrity/v1",
                 {
                     "Hashtag": "as:Hashtag",
-                    "sensitive": "as:sensitive"
+                    "sensitive": "as:sensitive",
+                    "toot": "http://joinmastodon.org/ns#",
+                    "Emoji": "toot:Emoji"
                 },
             ],
             "id": "https://server.example/.well-known/apgateway/did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/objects/11fa64ff-b5a3-47bf-b23d-22b360581c3f",