+++ mitra_activitypub/src/builders/offer_agreement.rs
@@ -152,6 +152,8 @@ mod tests {
                 {
                     "Hashtag": "as:Hashtag",
                     "sensitive": "as:sensitive",
+                    "toot": "http://joinmastodon.org/ns#",
+                    "Emoji": "toot:Emoji",
                     "vf": "https://w3id.org/valueflows/ont/vf#",
                     "om2": "http://www.ontology-of-units-of-measure.org/resource/om-2/",
                     "Proposal": "vf:Proposal",